+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Update API for table column sorting to pass models not properties to sorting blocks
 - Support table multi-column sort_property
+- Support table additional_sort_properties array
 - Fix issue with different sorting blocks not reseting each other on different table columns
 
 ## 0.2.4